update api v2, add /halts?height=_

+++ api/v2/service/service.go
@@ -26,9 +26,9 @@ func NewService(cdc *amino.Codec, blockchain *minter.Blockchain, client *rpc.Loc
 	return &Service{cdc: cdc, blockchain: blockchain, client: client, minterCfg: minterCfg, version: version, tmNode: node}
 }
 
-func (s *Service) getStateForHeight(height int32) (*state.State, error) {
+func (s *Service) getStateForHeight(height uint64) (*state.State, error) {
 	if height > 0 {
-		cState, err := s.blockchain.GetStateForHeight(uint64(height))
+		cState, err := s.blockchain.GetStateForHeight(height)
 		if err != nil {
 			return nil, err
 		}
